Move config setup code to its own class
