+++ lf-sample-repository-api-dotnet-srv.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.32510.428
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "lf-sample-repository-api-dotnet-srv", "lf-sample-repository-api-dotnet-srv.csproj", "{7EA539BB-4791-4C09-9485-573CDCD568FC}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{7EA539BB-4791-4C09-9485-573CDCD568FC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7EA539BB-4791-4C09-9485-573CDCD568FC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7EA539BB-4791-4C09-9485-573CDCD568FC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7EA539BB-4791-4C09-9485-573CDCD568FC}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {99C9E180-F315-44AB-999D-826E8B1C1B05}
+	EndGlobalSection
+EndGlobal
